공지사항 등록,삭제 안됨/후기 조금 건드림

+++ Sikhuchung/src/main/java/com/sikhuchung/domain/NoticeDTO.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class NoticeDTO {
-    int noticeNumber, noticeHit;
+    Long noticeNumber, noticeHit;
     String userId, noticeTitle, noticeContent, noticeDate, noticeUpdateId, noticeUpdateDate, noticeDelete;
 }
